+++ src/components/ens-registration/ConfirmContent/RegisterContent.tsx
@@ -1,7 +1,7 @@
 import lang from 'i18n-js';
 import React, { useCallback } from 'react';
 import { Switch } from 'react-native-gesture-handler';
-import { StepIndicator } from '../../../components/step-indicator';
+import StepIndicator from '../../../components/step-indicator/StepIndicator';
 import ButtonPressAnimation from '../../animations/ButtonPressAnimation';
 import {
   Box,
