--- src/components/large-countdown-clock/index.js
@@ -1,1 +1,0 @@
-export { default as LargeCountdownClock } from './LargeCountdownClock';
