--- src/components/step-indicator/index.js
@@ -1,1 +1,0 @@
-export { default as StepIndicator } from './StepIndicator';
