+++ src/hooks/index.js
@@ -72,6 +72,7 @@ export { default as useIsMounted } from './useIsMounted';
 export { default as useIsWalletEthZero } from './useIsWalletEthZero';
 export { default as useKeyboardHeight } from './useKeyboardHeight';
 export { default as useLoadAccountData } from './useLoadAccountData';
+export { default as useLoadAccountLateData } from './useLoadAccountLateData';
 export { default as useLoadGlobalEarlyData } from './useLoadGlobalEarlyData';
 export { default as useLoadGlobalLateData } from './useLoadGlobalLateData';
 export { default as useLongPressEvents } from './useLongPressEvents';
