+++ src/hooks/useLoadGlobalLateData.js
@@ -1,7 +1,5 @@
 import { useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { prefetchAccountENSDomains } from './useAccountENSDomains';
-import useAccountSettings from './useAccountSettings';
 import {
   getWalletBalances,
   WALLET_BALANCES_FROM_STORAGE,
@@ -21,8 +19,6 @@ export default function useLoadGlobalLateData() {
     ({ appState: { walletReady } }) => walletReady
   );
 
-  const { accountAddress } = useAccountSettings();
-
   const loadGlobalData = useCallback(async () => {
     if (!walletReady) {
       return false;
@@ -31,11 +27,10 @@ export default function useLoadGlobalLateData() {
     const promises = [];
     const p1 = loadWalletBalanceNamesToCache();
     const p2 = dispatch(nonceManagerLoadState());
-    const p3 = prefetchAccountENSDomains({ accountAddress });
-    promises.push(p1, p2, p3);
+    promises.push(p1, p2);
 
     return promiseUtils.PromiseAllWithFails(promises);
-  }, [accountAddress, dispatch, walletReady]);
+  }, [dispatch, walletReady]);
 
   return loadGlobalData;
 }
