+++ src/utils/index.ts
@@ -14,6 +14,14 @@ export { default as profileUtils } from './profileUtils';
 export { default as doesWalletsContainAddress } from './doesWalletsContainAddress';
 export { default as dimensionsPropType } from './dimensionsPropType';
 export { default as directionPropType } from './directionPropType';
+export {
+  getENSNFTAvatarUrl,
+  isENSNFTRecord,
+  normalizeENS,
+  parseENSNFTRecord,
+  stringifyENSNFTRecord,
+  validateENS,
+} from './ens';
 export { default as ethereumUtils } from './ethereumUtils';
 export { default as formatURLForDisplay } from './formatURLForDisplay';
 export { default as gasUtils } from './gas';
@@ -22,7 +30,6 @@ export { default as getDominantColorFromImage } from './getDominantColorFromImag
 export { default as getTokenMetadata } from './getTokenMetadata';
 export { default as getUrlForTrustIconFallback } from './getUrlForTrustIconFallback';
 export { default as haptics } from './haptics';
-export { default as isENSNFTRecord } from './isENSNFTRecord';
 export { default as isETH } from './isETH';
 export { default as isLowerCaseMatch } from './isLowerCaseMatch';
 export { default as isNewValueForObjectPaths } from './isNewValueForObjectPaths';
@@ -32,8 +39,6 @@ export { default as logger } from './logger';
 export { default as magicMemo } from './magicMemo';
 export { default as measureText } from './measureText';
 export { default as neverRerender } from './neverRerender';
-export { default as normalizeENS } from './normalizeENS';
-export { default as parseENSNFTRecord } from './parseENSNFTRecord';
 export { default as parseObjectToUrlQueryString } from './parseObjectToUrlQueryString';
 export { default as parseQueryParams } from './parseQueryParams';
 export { default as promiseUtils } from './promise';
@@ -44,7 +49,6 @@ export { default as sentryUtils } from './sentry';
 export { default as showActionSheetWithOptions } from './actionsheet';
 export { default as simplifyChartData } from './simplifyChartData';
 export { default as statusBar } from './statusBar';
-export { default as stringifyENSNFTRecord } from './stringifyENSNFTRecord';
 export { filterList } from './search';
 export {
   getFirstGrapheme,
@@ -52,7 +56,6 @@ export {
   removeLeadingZeros,
   sanitizeSeedPhrase,
 } from './formatters';
-export { default as validateENS } from './validateENS';
 export { default as watchingAlert } from './watchingAlert';
 export { default as withSpeed } from './withSpeed';
 export { default as CoinIcon } from './CoinIcons/CoinIcon';
