--- src/utils/isENSNFTRecord.ts
@@ -1,3 +1,0 @@
-export default function isENSNFTRecord(avatar: string) {
-  return avatar.includes('eip155:1');
-}
