+++ types/content-hash.d.ts
